part3 1345M 0.0232 beam=14 0.00908175  bash compare_v4d.pal.sh

+++ src/decoder/cuda-lattice-decoder.h
@@ -246,7 +246,7 @@ class CudaLatticeDecoder {
       return prev_tok_id >= ((uint32)1<<31);
     }
     HOST DEVICE inline uint32 GetPrevTokId() {
-      return prev_tok_id & ((1<<31) - 1);
+      return prev_tok_id & (((uint32)1<<31) - 1);
     }
   };
